added function that creates a copy of bg_node to the fg_node when bringing a bg process to the fg

+++ sources/builtins/ft_fg.c
@@ -6,7 +6,7 @@
 /*   By: mbarutel <mbarutel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/01 17:09:07 by mbarutel          #+#    #+#             */
-/*   Updated: 2023/02/01 17:23:13 by mbarutel         ###   ########.fr       */
+/*   Updated: 2023/02/02 13:17:16 by mbarutel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,7 +52,8 @@ int	ft_fg(t_shell *sh, char **cmd)
 	{
 		ft_print_dbl_array(*job->cmd);
 		ft_putchar('\n');
-		killpg(job->gpid, SIGCONT);
+		if (job->status == SUSPENDED)
+			killpg(job->gpid, SIGCONT);
 		waitpid(job->gpid, &status, WUNTRACED);
 	}
 	return (0);
